Fixes temp sensor lagging up ServoController Weirdly enough, the temp sensor was lagging up the servo controller.
This is a fix that is equally weird, but it works.
We simply split-off the temp sensor into its own process.

Thankfully, we still share the same node modules, so we don't have to
install them twice.

There should be no effect in the development and production envs,
as TypeScript will also compile our IsolatedDHTReader.ts even though
its not imported anywhere.